:recycle: rework `Macros.InnerInstancePool`
