+++ src/lib/macros/index.ts
@@ -4,5 +4,5 @@
  * @module
  */
 export * from './macro'
-export * from './inner-macros-pool'
 export * from './macro-workers'
+export * from './inner-observables-pool'
